% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf_tbl.R
\name{print.imf_tbl}
\alias{print.imf_tbl}
\title{Print an IMF data table}
\usage{
\method{print}{imf_tbl}(x, n = 10L, ...)
}
\arguments{
\item{x}{An \code{imf_tbl}, as returned by \code{\link[=imf_weo]{imf_weo()}}.}

\item{n}{Number of rows to print.}

\item{...}{Unused, for S3 compatibility.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an IMF data table
}
