% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readimf-package.R
\docType{package}
\name{readimf-package}
\alias{readimf}
\alias{readimf-package}
\title{readimf: Access International Monetary Fund Data via the 'SDMX' 3.0 API}
\description{
Download macroeconomic and financial statistics from the International Monetary Fund data portal (\url{https://data.imf.org}), served through an 'SDMX' 3.0 REST API. Provides a tidy interface to the Fund's flagship databases, beginning with the World Economic Outlook, and reads the compact 'SDMX-CSV' representation of each query for fast, dependency-light parsing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/charlescoverdale/readimf}
  \item Report bugs at \url{https://github.com/charlescoverdale/readimf/issues}
}

}
\author{
\strong{Maintainer}: Charles Coverdale \email{charlesfcoverdale@gmail.com}

}
\keyword{internal}
